# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mtaani Skills API - a marketplace connecting customers with verified TVET student Fundis
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  fundi: 'fundi',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  customer: 'customer',
  fundi: 'fundi',
  admin: 'admin',
} as const;

export interface UserInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 9 */
  phone: string;
  role: UserInputRole;
  email?: string;
}

export type FundiTvtLevel = typeof FundiTvtLevel[keyof typeof FundiTvtLevel];


export const FundiTvtLevel = {
  level_4: 'level_4',
  level_5: 'level_5',
  level_6: 'level_6',
} as const;

export interface Fundi {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userPhone?: string | null;
  skills: string[];
  tvtLevel: FundiTvtLevel;
  /** @nullable */
  certificateUrl?: string | null;
  /** @nullable */
  idNumber?: string | null;
  /** @nullable */
  bio?: string | null;
  hourlyRate: number;
  location: string;
  verified: boolean;
  available: boolean;
  /** @nullable */
  averageRating?: number | null;
  totalReviews?: number;
  completedJobs?: number;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export type FundiInputTvtLevel = typeof FundiInputTvtLevel[keyof typeof FundiInputTvtLevel];


export const FundiInputTvtLevel = {
  level_4: 'level_4',
  level_5: 'level_5',
  level_6: 'level_6',
} as const;

export interface FundiInput {
  userId: number;
  /** @minItems 1 */
  skills: string[];
  tvtLevel: FundiInputTvtLevel;
  certificateUrl?: string;
  idNumber?: string;
  bio?: string;
  /** @minimum 0 */
  hourlyRate: number;
  location: string;
  photoUrl?: string;
}

export interface FundiUpdate {
  skills?: string[];
  bio?: string;
  hourlyRate?: number;
  location?: string;
  available?: boolean;
  photoUrl?: string;
  certificateUrl?: string;
}

export interface FundiVerification {
  verified: boolean;
  rejectionReason?: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  in_progress: 'in_progress',
  completed: 'completed',
  disputed: 'disputed',
} as const;

export interface Job {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  fundiId: number;
  /** @nullable */
  fundiName?: string | null;
  description: string;
  location: string;
  agreedPrice: number;
  status: JobStatus;
  /** @nullable */
  scheduledDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  commissionAmount?: number | null;
  createdAt: string;
}

export interface JobInput {
  customerId: number;
  fundiId: number;
  /** @minLength 10 */
  description: string;
  location: string;
  /** @minimum 100 */
  agreedPrice: number;
  scheduledDate?: string;
  notes?: string;
}

export type JobUpdateStatus = typeof JobUpdateStatus[keyof typeof JobUpdateStatus];


export const JobUpdateStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
  in_progress: 'in_progress',
  completed: 'completed',
  disputed: 'disputed',
} as const;

export interface JobUpdate {
  status?: JobUpdateStatus;
  notes?: string;
  scheduledDate?: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  escrow: 'escrow',
  released: 'released',
  refunded: 'refunded',
} as const;

export interface Payment {
  id: number;
  jobId: number;
  amount: number;
  mpesaCode: string;
  status: PaymentStatus;
  /** @nullable */
  commissionAmount?: number | null;
  /** @nullable */
  fundiAmount?: number | null;
  createdAt: string;
}

export interface PaymentInput {
  jobId: number;
  /** @minimum 1 */
  amount: number;
  /** @minLength 6 */
  mpesaCode: string;
}

export interface Review {
  id: number;
  jobId: number;
  fundiId: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  jobId: number;
  fundiId: number;
  customerId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export interface DashboardStats {
  totalFundis: number;
  verifiedFundis: number;
  pendingVerifications: number;
  totalJobs: number;
  activeJobs: number;
  completedJobs: number;
  totalRevenue: number;
  commissionEarned: number;
  totalCustomers: number;
}

export type ListFundisParams = {
/**
 * Filter by skill category
 */
skill?: string;
/**
 * Filter by town / area
 */
location?: string;
/**
 * Only return verified fundis
 */
verified?: boolean;
/**
 * Only return available fundis
 */
available?: boolean;
};

export type ListJobsParams = {
/**
 * Filter by customer
 */
customerId?: number;
/**
 * Filter by fundi
 */
fundiId?: number;
/**
 * Filter by status
 */
status?: ListJobsStatus;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  in_progress: 'in_progress',
  completed: 'completed',
  disputed: 'disputed',
} as const;

export type ListPaymentsParams = {
jobId?: number;
};

export type ListReviewsParams = {
/**
 * Filter reviews by fundi
 */
fundiId?: number;
/**
 * Filter by job
 */
jobId?: number;
};

export type GetRecentJobsParams = {
limit?: number;
};

export type GetTopFundisParams = {
limit?: number;
};