# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mtaani Skills API - a marketplace connecting customers with verified TVET student Fundis
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fundis:
    get:
      operationId: listFundis
      tags: [fundis]
      summary: List / search fundis
      parameters:
        - name: skill
          in: query
          schema:
            type: string
          description: Filter by skill category
        - name: location
          in: query
          schema:
            type: string
          description: Filter by town / area
        - name: verified
          in: query
          schema:
            type: boolean
          description: Only return verified fundis
        - name: available
          in: query
          schema:
            type: boolean
          description: Only return available fundis
      responses:
        "200":
          description: List of fundis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fundi"
    post:
      operationId: createFundi
      tags: [fundis]
      summary: Register as a fundi (user becomes a fundi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundiInput"
      responses:
        "201":
          description: Fundi profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fundi"
        "400":
          description: Bad request
  /fundis/{id}:
    get:
      operationId: getFundi
      tags: [fundis]
      summary: Get fundi profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fundi profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fundi"
        "404":
          description: Not found
    patch:
      operationId: updateFundi
      tags: [fundis]
      summary: Update fundi profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundiUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fundi"
        "404":
          description: Not found
  /fundis/{id}/verify:
    patch:
      operationId: verifyFundi
      tags: [fundis]
      summary: Admin — approve/reject fundi verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundiVerification"
      responses:
        "200":
          description: Verification status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fundi"
        "404":
          description: Not found

  # ─── JOBS ─────────────────────────────────────────────────────────────────
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs (filter by role and status)
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
          description: Filter by customer
        - name: fundiId
          in: query
          schema:
            type: integer
          description: Filter by fundi
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, in_progress, completed, disputed]
          description: Filter by status
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Customer creates a job booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Bad request
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job status (accept, reject, complete, dispute)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
  /jobs/{id}/release-payment:
    post:
      operationId: releaseJobPayment
      tags: [jobs]
      summary: Customer or admin releases escrow payment to fundi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Cannot release payment
        "404":
          description: Not found

  # ─── PAYMENTS ─────────────────────────────────────────────────────────────
...
  /dashboard/recent-jobs:
    get:
      operationId: getRecentJobs
      tags: [dashboard]
      summary: Most recent jobs across the platform
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /dashboard/top-fundis:
    get:
      operationId: getTopFundis
      tags: [dashboard]
      summary: Top-rated verified fundis
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Top fundis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fundi"
...
components:
  schemas:
...
    Fundi:
      type: object
      required: [id, userId, skills, tvtLevel, hourlyRate, location, verified, available, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        tvtLevel:
          type: string
          enum: [level_4, level_5, level_6]
        certificateUrl:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        hourlyRate:
          type: number
        location:
          type: string
        verified:
          type: boolean
        available:
          type: boolean
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        completedJobs:
          type: integer
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    FundiInput:
      type: object
      required: [userId, skills, tvtLevel, hourlyRate, location]
      properties:
        userId:
          type: integer
        skills:
          type: array
          items:
            type: string
          minItems: 1
        tvtLevel:
          type: string
          enum: [level_4, level_5, level_6]
        certificateUrl:
          type: string
        idNumber:
          type: string
        bio:
          type: string
        hourlyRate:
          type: number
          minimum: 0
        location:
          type: string
        photoUrl:
          type: string
    FundiUpdate:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        hourlyRate:
          type: number
        location:
          type: string
        available:
          type: boolean
        photoUrl:
          type: string
        certificateUrl:
          type: string
    FundiVerification:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
        rejectionReason:
          type: string

    # ─── JOB ─────────────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, customerId, fundiId, description, location, agreedPrice, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        fundiId:
          type: integer
        fundiName:
          type: ["string", "null"]
        description:
          type: string
        location:
          type: string
        agreedPrice:
          type: number
        status:
          type: string
          enum: [pending, accepted, rejected, in_progress, completed, disputed]
        scheduledDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        commissionAmount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    JobInput:
      type: object
      required: [customerId, fundiId, description, location, agreedPrice]
      properties:
        customerId:
          type: integer
        fundiId:
          type: integer
        description:
          type: string
          minLength: 10
        location:
          type: string
        agreedPrice:
          type: number
          minimum: 100
        scheduledDate:
          type: string
        notes:
          type: string
    JobUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected, in_progress, completed, disputed]
        notes:
          type: string
        scheduledDate:
          type: string

    # ─── PAYMENT ─────────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, jobId, amount, mpesaCode, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        amount:
          type: number
        mpesaCode:
          type: string
        status:
          type: string
          enum: [escrow, released, refunded]
        commissionAmount:
          type: ["number", "null"]
        fundiAmount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...