# lib/api-zod/src/generated/types/fundiInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mtaani Skills API - a marketplace connecting customers with verified TVET student Fundis
 * OpenAPI spec version: 0.1.0
 */
import type { FundiInputTvtLevel } from './fundiInputTvtLevel';

export interface FundiInput {
  userId: number;
  /** @minItems 1 */
  skills: string[];
  tvtLevel: FundiInputTvtLevel;
  certificateUrl?: string;
  idNumber?: string;
  bio?: string;
  /** @minimum 0 */
  hourlyRate: number;
  location: string;
  photoUrl?: string;
}