# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mtaani Skills API - a marketplace connecting customers with verified TVET student Fundis
 * OpenAPI spec version: 0.1.0
 */
import type { JobUpdateStatus } from './jobUpdateStatus';

export interface JobUpdate {
  status?: JobUpdateStatus;
  notes?: string;
  scheduledDate?: string;
}