# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mtaani Skills API - a marketplace connecting customers with verified TVET student Fundis
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentStatus } from './paymentStatus';

export interface Payment {
  id: number;
  jobId: number;
  amount: number;
  mpesaCode: string;
  status: PaymentStatus;
  /** @nullable */
  commissionAmount?: number | null;
  /** @nullable */
  fundiAmount?: number | null;
  createdAt: Date;
}